Bug 1098970 - Part 6: Fix the usage of GonkNativeWindow in Omx. r=sotaro
Use new APIs for GonkNativeWindowLL.

UltraBlame original commit: 7e39cc7d3cf6505d0791386a6769a7fd82c9ce55

+++ dom/media/omx/MediaOmxReader.cpp
@@ -112,7 +112,7 @@ private:
       mOffset += length;
     }
 
-    if (mOffset < mFullLength) {
+    if (static_cast<uint64_t>(mOffset) < mFullLength) {
